added null check

+++ drip-api/src/main/java/nl/uva/sne/drip/api/service/ConfigurationService.java
@@ -89,6 +89,9 @@ public class ConfigurationService {
     @PostAuthorize("(returnObject.owner == authentication.name) or (hasRole('ROLE_ADMIN'))")
     public ConfigurationRepresentation delete(String id) {
         ConfigurationRepresentation tr = dao.findOne(id);
+        if(tr == null){
+            throw new NotFoundException();
+        }
         dao.delete(tr);
         return tr;
     }
